refactor: move math functions to a cmath.py file
